Fixes CDS frame
It was expected that all exon CDS are in the first reading frame (0). In GRCm38.87, there are 3570 proteins with the first exon in another frame (GTF column 8 - genomic phase (for CDS features) )  (e.g. ENSMUST00000149214, ENSMUST00000190155). Those proteins were thus incorrectly translated and had many stop in the resulting sequence. 

Resolution:
Added frame a check for the first exon of the CDS. It truncates unnecessary nucleotides from the start.